Разбил reviews.js на модули
Common JS
