+++ src/getReviewElement.js
@@ -1,0 +1,50 @@
+'use strict';
+
+window.reviewsContainer = document.querySelector('.reviews-list');
+
+var templateElement = document.querySelector('template');
+var IMAGE_LOAD_TIMEOUT = 10000;
+var elementToClone;
+
+if ('content' in templateElement) {
+  elementToClone = templateElement.content.querySelector('.review');
+} else {
+  elementToClone = templateElement.querySelector('.review');
+}
+
+ var getReviewElement = function(data, container) {
+  var element = elementToClone.cloneNode(true);
+
+  var ratingStar = element.querySelector('.review-rating');
+  ratingStar.style.width = data.rating * 30 + 'px';
+
+  element.querySelector('.review-text').textContent = data.description;
+
+  var userPhoto = new Image(124, 124);
+  var imageLoadTimeout;
+
+  userPhoto.onerror = function() {
+    element.classList.add('review-load-failure');
+  };
+
+  userPhoto.onload = function() {
+    clearTimeout(imageLoadTimeout);
+  };
+
+  userPhoto.src = data.author.picture;
+  element.querySelector('.review-author').src = userPhoto.src;
+  element.querySelector('.review-author').title = data.author.name;
+  element.querySelector('.review-author').alt = data.author.name;
+
+  imageLoadTimeout = setTimeout(function() {
+    userPhoto.src = '';
+    element.classList.add('review-load-failure');
+  }, IMAGE_LOAD_TIMEOUT);
+
+  container.appendChild(element);
+  return element;
+};
+
+module.exports = {
+  build: getReviewElement
+};
